Created mother class for FFT convolution filters and modified ramp and scatter glare accordingly

+++ code/rtkCudaFFTRampImageFilter.h
@@ -47,7 +47,7 @@ public:
   typedef itk::CudaImage< std::complex<float>, 3 >                   CudaFFTOutputImageType;
   typedef CudaFFTOutputImageType::Pointer                            CudaFFTOutputImagePointer;
   typedef CudaFFTRampImageFilter                                     Self;
-  typedef FFTRampImageFilter< CudaImageType, CudaImageType, double > Superclass;
+  typedef FFTRampImageFilter< CudaImageType, CudaImageType, float >  Superclass;
   typedef itk::SmartPointer<Self>                                    Pointer;
   typedef itk::SmartPointer<const Self>                              ConstPointer;
